pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-docs-summarizer"
version = "0.1.0"
description = "A tool to crawl and summarize MCP documentation using deepseek-chat"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing",
]

dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "python-dotenv>=1.1.1",
    "openai>=1.93.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.1",
]

[project.scripts]
mcp-summarizer = "main:main"